refactor: 💡 hearing outcome table
