+++ src/components/HearingOutcome/utils.ts
@@ -1,0 +1,11 @@
+import { Offence } from "@moj-bichard7-developers/bichard7-next-core/build/src/types/AnnotatedHearingOutcome"
+
+export const getOffenceCode = (offence: Offence): string => {
+  if (offence.CriminalProsecutionReference.OffenceReason?.__type === "LocalOffenceReason") {
+    return offence.CriminalProsecutionReference.OffenceReason.LocalOffenceCode.OffenceCode
+  }
+  if (offence.CriminalProsecutionReference.OffenceReason?.__type === "NationalOffenceReason") {
+    return offence.CriminalProsecutionReference.OffenceReason?.OffenceCode.FullCode
+  }
+  return ""
+}
