+++ src/services/amendCourtCase/applyAmendmentsToAho.ts
@@ -46,6 +46,7 @@ const selectKey = (aho: AnnotatedHearingOutcome) => (key: string, value: Amendme
       amendCourtPNCIdentifier(value as string, aho)
       break
     case "resultVariableText":
+      console.log("v", value)
       amendResultVariableText(value as UpdatedOffenceResult[], aho)
       break
     case "courtReference":
